Safari-609.1.11.1/LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-volume-styles-expected.txt

Testing the MacOSInlineMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.volumeSlider.parent.element.parentNode became mediaControls.element
PASS volumeSliderContainerStyle.position is "absolute"
PASS volumeSliderContainerStyle.left is "566px"
PASS volumeSliderContainerStyle.top is "226px"
PASS volumeSliderContainerStyle.width is "94px"
PASS volumeSliderContainerStyle.height is "31px"
PASS volumeSliderContainerStyle.transform is "matrix(6.123233995736766e-17, -1, 1, 6.123233995736766e-17, -30, -26.000000000000004)"
PASS volumeSliderStyle.marginLeft is "6px"
PASS volumeSliderStyle.width is "60px"
PASS volumeSliderStyle.height is "16px"

PASS successfullyParsed is true

TEST COMPLETE
